Removes Recommendations and Employee modules, consolidates hotel recommendations
Eliminates the Recommendations and Employee modules, including models, DTOs, controllers, services, and mappings, to streamline the codebase.
Moves user-based hotel recommendation logic into the hotel domain, exposing recommendations via the hotel API and updating the frontend call accordingly.
Simplifies backend interfaces and service registrations by removing unused abstractions and dependencies.
Enhances maintainability by consolidating recommendation logic and reducing code redundancy.
