+++ backend/eBooking/API/Interfaces/IBaseService.cs
@@ -1,20 +1,7 @@
 ﻿using API.DTOs;
-using API.Models;
 
 namespace API.Interfaces
 {
-    public interface IBaseService<T> where T : BaseEntity
-    {
-        Task<T?> GetByIdAsync(int id);
-        Task<IEnumerable<T>> GetAllAsync();
-        Task<IEnumerable<T>> GetAllAsync(int pageNumber, int pageSize);
-        Task<T> AddAsync(T entity);
-        Task<T> UpdateAsync(T entity);
-        Task<bool> DeleteAsync(int id);
-        Task<bool> ExistsAsync(int id);
-        Task<int> CountAsync();
-    }
-
     public interface IBaseService<TDto, TCreateDto, TUpdateDto>
         where TDto : BaseEntityDto
         where TCreateDto : CreateBaseEntityDto
@@ -29,5 +16,4 @@ namespace API.Interfaces
         Task<bool> ExistsAsync(int id);
         Task<int> CountAsync();
     }
-
 }
