--- backend/eBooking/API/Services/BookingCreatedConsumer.cs
@@ -1,3 +1,0 @@
-// premješteno u API.Messaging.Consumers
-
-
