--- backend/eBooking/API/Services/BookingUpdatedConsumer.cs
@@ -1,3 +1,0 @@
-// premješteno u API.Messaging.Consumers
-
-
