--- backend/eBooking/API/Services/NotificationCreatedConsumer.cs
@@ -1,3 +1,0 @@
-// premješteno u API.Messaging.Consumers
-
-
