--- backend/eBooking/API/Services/UpcomingCheckInReminderConsumer.cs
@@ -1,3 +1,0 @@
-// premješteno u API.Messaging.Consumers
-
-
